sysctl takes a 'size_t' variable for the size parameter
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-4-patches@19829 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/hardware/memory/memory_netbsd.c
@@ -30,7 +30,7 @@ int netsnmp_mem_arch_load( netsnmp_cache *cache, void *magic ) {
     long           pagesize;
 
     struct uvmexp  uvmexp;
-    int            uvmexp_size  = sizeof(uvmexp);
+    size_t         uvmexp_size  = sizeof(uvmexp);
     int            uvmexp_mib[] = { CTL_VM, VM_UVMEXP };
 
     struct vmtotal total;
